refactor: use /v2 path suffix
