+++ elixir/lib/portal_api/client/v2/socket.ex
@@ -1,0 +1,11 @@
+defmodule PortalAPI.Client.V2.Socket do
+  use Phoenix.Socket
+
+  channel "client", PortalAPI.Client.V2.Channel
+
+  @impl true
+  defdelegate connect(attrs, socket, connect_info), to: PortalAPI.Client.Socket
+
+  @impl true
+  defdelegate id(socket), to: PortalAPI.Client.Socket
+end
