+++ elixir/lib/portal_api/endpoint.ex
@@ -41,7 +41,27 @@ defmodule PortalAPI.Endpoint do
     longpoll: false,
     drainer: []
 
+  socket "/gateway/v2", PortalAPI.Gateway.V2.Socket,
+    websocket: [
+      check_origin: :conn,
+      connect_info: [:trace_context_headers, :user_agent, :peer_data, :x_headers],
+      error_handler: {PortalAPI.Sockets, :handle_error, []},
+      timeout: :timer.seconds(37)
+    ],
+    longpoll: false,
+    drainer: []
+
   socket "/client", PortalAPI.Client.Socket,
+    websocket: [
+      check_origin: :conn,
+      connect_info: [:trace_context_headers, :user_agent, :peer_data, :x_headers],
+      error_handler: {PortalAPI.Sockets, :handle_error, []},
+      timeout: :timer.seconds(37)
+    ],
+    longpoll: false,
+    drainer: []
+
+  socket "/client/v2", PortalAPI.Client.V2.Socket,
     websocket: [
       check_origin: :conn,
       connect_info: [:trace_context_headers, :user_agent, :peer_data, :x_headers],
